separate loadGraph
